Halloween update
Auto use Halloween loading screens.
Improve dgVoodoo settings initialization.

+++ Assets/mainData/Assets/Scripts/PreloadingSplashScreen.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class PreloadingSplashScreen : MonoBehaviour
 { System.Security.Cryptography.MD5CryptoServiceProvider md = null;
-        
+		
 	public GUIStyle presentsStyle;
 
 	private Texture2D splashTexture;
@@ -55,46 +55,57 @@ public class PreloadingSplashScreen : MonoBehaviour
 
 
 		string filePath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), "dgVoodoo/dgVoodoo.conf");
+		string filePathDLL = Path.Combine(Path.Combine(Application.dataPath, ".."), "d3d9.dll");
 
-	    string searchString = "dgVoodooWatermark                   = true";
+		string searchString = "dgVoodooWatermark                   = true";
 		string vsyncSearchString = "ForceVerticalSync                   = false";
-    	string replacementString = "dgVoodooWatermark                   = false";
+		string replacementString = "dgVoodooWatermark                   = false";
 		string vsyncReplacementString = "ForceVerticalSync                   = true";
 		string fullconf = "[DirectX]\nForceVerticalSync                   = true\ndgVoodooWatermark                   = false";
-		if (File.Exists(filePath))
-        {
-			try
+		// if (File.Exists(filePathDLL) && !File.Exists(filePathDXGIDLL))
+		if (File.Exists(filePathDLL))
+		{
+			if (File.Exists(filePath))
 			{
-				string fileContents = File.ReadAllText(filePath);
-				fileContents = fileContents.Replace(searchString, replacementString);
-				fileContents = fileContents.Replace(vsyncSearchString, vsyncReplacementString);
-				File.WriteAllText(filePath, fileContents);
-			}
-			catch (Exception e)
-			{
-				CspUtils.DebugLogError("Error replacing string in dgVoodoo config: " + e.Message);
-			}
-        }
-        else
-        {
-            // File doesn't exist, so create it and write the replacement string
-              try
+				try
 				{
-					// Ensure the parent directories exist before creating the file
-					Directory.CreateDirectory(Path.GetDirectoryName(filePath));
-					// Create a new file
-					File.WriteAllText(filePath, fullconf);
-					// Debug.Log("File created with replacement string.");
+					string fileContents = File.ReadAllText(filePath);
+					fileContents = fileContents.Replace(searchString, replacementString);
+					fileContents = fileContents.Replace(vsyncSearchString, vsyncReplacementString);
+					File.WriteAllText(filePath, fileContents);
 				}
 				catch (Exception e)
 				{
-					Debug.LogError("Error creating dgVoodoo config file: " + e.Message);
+					CspUtils.DebugLogError("Error replacing string in dgVoodoo config: " + e.Message);
 				}
-        }
+			}
+			else
+			{
+				// File doesn't exist, so create it and write the replacement string
+				try
+					{
+						// Ensure the parent directories exist before creating the file
+						Directory.CreateDirectory(Path.GetDirectoryName(filePath));
+						// Create a new file
+						File.WriteAllText(filePath, fullconf);
+						// Debug.Log("File created with replacement string.");
+					}
+					catch (Exception e)
+					{
+						CspUtils.DebugLogError("Error creating dgVoodoo config file: " + e.Message);
+					}
+			}
+		}
 
-		System.Random random = new System.Random();
-		splashTexture = (Texture2D)Resources.Load("GUI/loading/background/" + random.Next(0, 7).ToString());
-		SHSULogo = (Texture2D)Resources.Load("GUI/loading/logos/SHSU");
+		// System.Random random = new System.Random();
+		if (DateTime.Now.ToString("MM") == "10") {
+				CspUtils.maxFilename = CspUtils.maxFilenameHalloween;
+				CspUtils.halloweenSuffix = "_halloween";
+				CspUtils.halloweenPathSuffix = "halloween/";
+			}
+		int randomBackgroundNumber = new System.Random().Next(0, CspUtils.maxFilename);
+		splashTexture = (Texture2D)Resources.Load("GUI/loading/background/" + CspUtils.halloweenPathSuffix + randomBackgroundNumber);
+		SHSULogo = (Texture2D)Resources.Load("GUI/loading/logos/SHSU" + CspUtils.halloweenSuffix);
 		if (splashTexture != null)
 		{
 			width = Screen.width;
